refactor: move RunnerStatusEnricher to RunnerBarCore/Runner (#925)
- Delete Sources/RunnerBar/Runner/RunnerStatusEnricher.swift
- Add Sources/RunnerBarCore/Runner/RunnerStatusEnricher.swift with public API
- Remove import RunnerBarCore (can't import yourself)
- No import changes needed in callers (covered by @_exported in Exports.swift)
